projects/34/requirements.txt
flask>=2.0.1
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
scikit-image>=0.22.0
Pillow>=10.0.0
requests>=2.31.0
torchxrayvision>=0.0.39
werkzeug>=2.0.3 